src/requirements.txt
# SparkYouTube AI Blog System - Requirements
# Core AI and processing dependencies
google-generativeai>=0.8.0
python-dotenv>=1.0.0

# YouTube transcript extraction
youtube-transcript-api>=0.6.0

# Video processing and metadata
yt-dlp>=2024.8.6

# Language detection
langdetect>=1.0.9

# HTTP requests and async support
aiohttp>=3.9.0
aiofiles>=23.0.0

# Text processing and utilities
python-slugify>=8.0.0
pyyaml>=6.0.0

# Optional: For enhanced functionality
# googletrans==4.0.0rc1  # For translation features
# openai-whisper>=20231117  # For audio transcription (if needed)

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0
